fix: resolve config package shadowing and remap hikvision-sim to port 9011
- Move Settings class from config.py into config/__init__.py so that
  the config/ package correctly exposes `settings` alongside zone_matrix
- Delete the now-unreachable config.py (package always wins over module)
- Change hikvision-sim host port from 9001 → 9011 (9001 is MinIO console)
